Fail to read/write if register number is not parsed correctly
It's possible that a user will provide a register number in a format
that strtoul() can't parse. strtoul() will return 0 if it fails to parse
a value. The msr-tools will then silently read/write to MSR 0x0 and
succeed. At best this may confuse the user, or at worst this will
unintentionally write bad data to the system.

Check that the register number parsing has succeeded. If it fails, then
inform the user, give a suggestion, and exit.

Signed-off-by: Yazen Ghannam <yazen.ghannam@amd.com>